+++ src/main/java/io/github/ImpactDevelopment/installer/setting/settings/MinecraftVersionSetting.java
@@ -39,6 +39,11 @@ public enum MinecraftVersionSetting implements ChoiceSetting<String> {
     }
 
     @Override
+    public boolean validSetting(InstallationConfig config, String value) {
+        return true;
+    }
+
+    @Override
     public String toString() {
         return getClass().getSimpleName();
     }